+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/clannad/ProfileController.java
@@ -1,13 +1,16 @@
 package com.dwarfeng.familyhelper.webapi.node.controller.v1.clannad;
 
+import com.dwarfeng.familyhelper.clannad.sdk.bean.dto.WebInputProfileUpdateInfo;
 import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonProfile;
-import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.WebInputProfile;
 import com.dwarfeng.familyhelper.clannad.stack.bean.entity.Profile;
 import com.dwarfeng.familyhelper.webapi.sdk.bean.disp.clannad.FastJsonDispProfile;
+import com.dwarfeng.familyhelper.webapi.sdk.cna.ValidateList;
 import com.dwarfeng.familyhelper.webapi.stack.bean.disp.clannad.DispProfile;
+import com.dwarfeng.familyhelper.webapi.stack.handler.system.TokenHandler;
 import com.dwarfeng.familyhelper.webapi.stack.service.clannad.ProfileResponseService;
 import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
 import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.bean.key.WebInputStringIdKey;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.http.BindingCheck;
 import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
@@ -18,6 +21,7 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.stream.Collectors;
 
 /**
  * 个人信息控制器。
@@ -30,13 +34,18 @@ import javax.servlet.http.HttpServletRequest;
 public class ProfileController {
 
     private final ProfileResponseService service;
+
     private final ServiceExceptionMapper sem;
 
+    private final TokenHandler tokenHandler;
+
+
     public ProfileController(
-            ProfileResponseService service, ServiceExceptionMapper sem
+            ProfileResponseService service, ServiceExceptionMapper sem, TokenHandler tokenHandler
     ) {
         this.service = service;
         this.sem = sem;
+        this.tokenHandler = tokenHandler;
     }
 
     @GetMapping("/profile/{id}/exists")
@@ -77,16 +86,37 @@ public class ProfileController {
         }
     }
 
-    @PatchMapping("/profile")
+    @PatchMapping("/profile/update")
     @BehaviorAnalyse
     @BindingCheck
     @LoginRequired
     public FastJsonResponseData<Object> updateProfile(
             HttpServletRequest request,
-            @RequestBody @Validated WebInputProfile webInputProfile, BindingResult bindingResult
+            @RequestBody @Validated WebInputProfileUpdateInfo webInputProfileUpdateInfo, BindingResult bindingResult
     ) {
         try {
-            service.updateProfile(WebInputProfile.toStackBean(webInputProfile));
+            StringIdKey accountKey = tokenHandler.getAccountKey(request);
+            service.updateProfile(accountKey, WebInputProfileUpdateInfo.toStackBean(webInputProfileUpdateInfo));
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    @PostMapping("/profile/reset-guest-permission")
+    @BehaviorAnalyse
+    @BindingCheck
+    @LoginRequired
+    public FastJsonResponseData<Object> resetGuestPermission(
+            HttpServletRequest request,
+            @RequestBody @Validated ValidateList<WebInputStringIdKey> guestKeys, BindingResult bindingResult
+    ) {
+        try {
+            StringIdKey accountKey = tokenHandler.getAccountKey(request);
+            service.resetGuestPermission(
+                    accountKey,
+                    guestKeys.stream().map(WebInputStringIdKey::toStackBean).collect(Collectors.toList())
+            );
             return FastJsonResponseData.of(ResponseDataUtil.good(null));
         } catch (Exception e) {
             return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));