+++ familyhelper-webapi-stack/src/main/java/com/dwarfeng/familyhelper/webapi/stack/service/clannad/ProfileResponseService.java
@@ -1,10 +1,13 @@
 package com.dwarfeng.familyhelper.webapi.stack.service.clannad;
 
+import com.dwarfeng.familyhelper.clannad.stack.bean.dto.ProfileUpdateInfo;
 import com.dwarfeng.familyhelper.clannad.stack.bean.entity.Profile;
 import com.dwarfeng.familyhelper.webapi.stack.bean.disp.clannad.DispProfile;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.service.Service;
+
+import java.util.Collection;
 
 /**
  * 银行卡响应服务。
@@ -20,5 +23,11 @@ public interface ProfileResponseService extends Service {
 
     DispProfile getDisp(StringIdKey key) throws ServiceException;
 
-    void updateProfile(Profile profile) throws ServiceException;
+    void updateProfile(StringIdKey userKey, ProfileUpdateInfo profileUpdateInfo) throws ServiceException;
+
+    void addGuestPermission(StringIdKey ownerUserKey, StringIdKey guestUserKey) throws ServiceException;
+
+    void removeGuestPermission(StringIdKey ownerUserKey, StringIdKey guestUserKey) throws ServiceException;
+
+    void resetGuestPermission(StringIdKey ownerUserKey, Collection<StringIdKey> guestUserKeys) throws ServiceException;
 }